Repackage UI classes and fxml files
