+++ src/main/java/seedu/guilttrip/ui/autoexpense/AutoExpensesCard.java
@@ -1,4 +1,4 @@
-package seedu.guilttrip.ui;
+package seedu.guilttrip.ui.autoexpense;
 
 import java.util.Comparator;
 
@@ -7,14 +7,15 @@ import javafx.scene.control.Label;
 import javafx.scene.layout.FlowPane;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Region;
-import seedu.guilttrip.model.entry.Entry;
+import seedu.guilttrip.model.entry.AutoExpense;
+import seedu.guilttrip.ui.UiPart;
 
 /**
  * An UI component that displays information of a {@code Person}.
  */
-public class EntryCard extends UiPart<Region> {
+public class AutoExpensesCard extends UiPart<Region> {
 
-    private static final String FXML = "EntryListCard.fxml";
+    private static final String FXML = "/autoexpense/AutoExpenseListCard.fxml";
 
     /**
      * Note: Certain keywords such as "location" and "resources" are reserved keywords in JavaFX.
@@ -24,7 +25,7 @@ public class EntryCard extends UiPart<Region> {
      * @see <a href="https://github.com/se-edu/addressbook-level4/issues/336">The issue on GuiltTrip level 4</a>
      */
 
-    public final Entry entry;
+    public final AutoExpense autoExpense;
 
     @FXML
     private HBox cardPane;
@@ -33,7 +34,9 @@ public class EntryCard extends UiPart<Region> {
     @FXML
     private Label id;
     @FXML
-    private Label date;
+    private Label freq;
+    @FXML
+    private Label nextTime;
     @FXML
     private Label amt;
     @FXML
@@ -41,18 +44,20 @@ public class EntryCard extends UiPart<Region> {
     @FXML
     private FlowPane tags;
 
-    public EntryCard(Entry entry, int displayedIndex) {
+    public AutoExpensesCard(AutoExpense autoExpense, int displayedIndex) {
         super(FXML);
-        this.entry = entry;
+        this.autoExpense = autoExpense;
         id.setText(displayedIndex + ". ");
 
-        String descWithType = entry.getDesc().fullDesc;
-        desc.setText(descWithType);
-        date.setText(entry.getDate().toString());
-        amt.setText("$" + entry.getAmount().value);
-        category.setText(entry.getCategory().categoryName);
+        String fullDesc = autoExpense.getDesc().fullDesc;
+        desc.setText(fullDesc);
+        freq.setText("[" + autoExpense.getFrequency().toString() + "] ");
+        //lastTime.setText(autoExpense.getLastTime().toString());
+        nextTime.setText("Next: " + autoExpense.getNextTime().toString());
+        amt.setText("$" + autoExpense.getAmount().toString());
+        category.setText(autoExpense.getCategory().toString());
 
-        entry.getTags().stream()
+        autoExpense.getTags().stream()
                 .sorted(Comparator.comparing(tag -> tag.tagName))
                 .forEach(tag -> tags.getChildren().add(new Label(tag.tagName)));
     }
@@ -65,13 +70,14 @@ public class EntryCard extends UiPart<Region> {
         }
 
         // instanceof handles nulls
-        if (!(other instanceof EntryCard)) {
+        if (!(other instanceof AutoExpensesCard)) {
             return false;
         }
 
         // state check
-        EntryCard card = (EntryCard) other;
+        AutoExpensesCard card = (AutoExpensesCard) other;
         return id.getText().equals(card.id.getText())
-                && entry.equals(card.entry);
+                && autoExpense.equals(card.autoExpense);
     }
 }
+
