+++ src/main/java/seedu/guilttrip/ui/autoexpense/AutoExpensesPanel.java
@@ -1,4 +1,4 @@
-package seedu.guilttrip.ui;
+package seedu.guilttrip.ui.autoexpense;
 
 import java.util.logging.Logger;
 
@@ -9,12 +9,13 @@ import javafx.scene.control.ListView;
 import javafx.scene.layout.Region;
 import seedu.guilttrip.commons.core.LogsCenter;
 import seedu.guilttrip.model.entry.AutoExpense;
+import seedu.guilttrip.ui.UiPart;
 
 /**
  * Side panel for budgets.
  */
 public class AutoExpensesPanel extends UiPart<Region> {
-    private static final String FXML = "AutoExpenseListPanel.fxml";
+    private static final String FXML = "/autoexpense/AutoExpenseListPanel.fxml";
     private final Logger logger = LogsCenter.getLogger(AutoExpensesPanel.class);
 
     @FXML
