+++ src/main/java/seedu/guilttrip/ui/expense/ExpenseCard.java
@@ -1,21 +1,21 @@
-package seedu.guilttrip.ui;
+package seedu.guilttrip.ui.expense;
 
 import java.util.Comparator;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
-import javafx.scene.control.ProgressBar;
 import javafx.scene.layout.FlowPane;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Region;
-import seedu.guilttrip.model.entry.Budget;
+import seedu.guilttrip.model.entry.Expense;
+import seedu.guilttrip.ui.UiPart;
 
 /**
  * An UI component that displays information of a {@code Person}.
  */
-public class BudgetCard extends UiPart<Region> {
+public class ExpenseCard extends UiPart<Region> {
 
-    private static final String FXML = "BudgetListCard.fxml";
+    private static final String FXML = "/expense/ExpenseCard.fxml";
 
     /**
      * Note: Certain keywords such as "location" and "resources" are reserved keywords in JavaFX.
@@ -25,7 +25,7 @@ public class BudgetCard extends UiPart<Region> {
      * @see <a href="https://github.com/se-edu/addressbook-level4/issues/336">The issue on GuiltTrip level 4</a>
      */
 
-    public final Budget budget;
+    public final Expense expense;
 
     @FXML
     private HBox cardPane;
@@ -36,35 +36,24 @@ public class BudgetCard extends UiPart<Region> {
     @FXML
     private Label date;
     @FXML
-    private Label left;
-    @FXML
-    private ProgressBar progressBar;
-    @FXML
-    private Label total;
+    private Label amt;
     @FXML
     private Label category;
     @FXML
     private FlowPane tags;
 
-    public BudgetCard(Budget budget, int displayedIndex) {
+    public ExpenseCard(Expense expense, int displayedIndex) {
         super(FXML);
-        this.budget = budget;
+        this.expense = expense;
         id.setText(displayedIndex + ". ");
 
-        String fullDesc = budget.getDesc().fullDesc;
-        desc.setText(fullDesc);
-        date.setText(budget.getDate().toString() + " - " + budget.getEndDate().toString());
-        double leftAmount = budget.getAmount().value - budget.getSpent().value;
-        double totalAmount = budget.getAmount().value;
-        //left.setText("left: $" + leftAmount + " out of: $" + budget.getAmount().value);
-        left.setText("$" + leftAmount);
-        total.setText("$" + totalAmount);
+        String descWithType = expense.getDesc().fullDesc;
+        desc.setText(descWithType);
+        date.setText(expense.getDate().toString());
+        amt.setText("$" + expense.getAmount().value);
+        category.setText(expense.getCategory().categoryName);
 
-        progressBar.setProgress(leftAmount / totalAmount);
-
-        category.setText(budget.getCategory().toString());
-
-        budget.getTags().stream()
+        expense.getTags().stream()
                 .sorted(Comparator.comparing(tag -> tag.tagName))
                 .forEach(tag -> tags.getChildren().add(new Label(tag.tagName)));
     }
@@ -77,14 +66,13 @@ public class BudgetCard extends UiPart<Region> {
         }
 
         // instanceof handles nulls
-        if (!(other instanceof BudgetCard)) {
+        if (!(other instanceof ExpenseCard)) {
             return false;
         }
 
         // state check
-        BudgetCard card = (BudgetCard) other;
+        ExpenseCard card = (ExpenseCard) other;
         return id.getText().equals(card.id.getText())
-                && budget.equals(card.budget);
+                && expense.equals(card.expense);
     }
 }
-
