+++ src/main/java/seedu/guilttrip/ui/wishlist/WishListCard.java
@@ -1,4 +1,4 @@
-package seedu.guilttrip.ui;
+package seedu.guilttrip.ui.wishlist;
 
 import java.util.Comparator;
 
@@ -7,14 +7,16 @@ import javafx.scene.control.Label;
 import javafx.scene.layout.FlowPane;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Region;
-import seedu.guilttrip.model.entry.Income;
+import seedu.guilttrip.model.entry.Wish;
+import seedu.guilttrip.ui.UiPart;
+import seedu.guilttrip.ui.entry.EntryCard;
 
 /**
  * An UI component that displays information of a {@code Person}.
  */
-public class IncomeCard extends UiPart<Region> {
+public class WishListCard extends UiPart<Region> {
 
-    private static final String FXML = "IncomeCard.fxml";
+    private static final String FXML = "/wishlist/WishListCard.fxml";
 
     /**
      * Note: Certain keywords such as "location" and "resources" are reserved keywords in JavaFX.
@@ -24,7 +26,7 @@ public class IncomeCard extends UiPart<Region> {
      * @see <a href="https://github.com/se-edu/addressbook-level4/issues/336">The issue on GuiltTrip level 4</a>
      */
 
-    public final Income income;
+    public final Wish wish;
 
     @FXML
     private HBox cardPane;
@@ -41,18 +43,18 @@ public class IncomeCard extends UiPart<Region> {
     @FXML
     private FlowPane tags;
 
-    public IncomeCard(Income income, int displayedIndex) {
+    public WishListCard(Wish wish, int displayedIndex) {
         super(FXML);
-        this.income = income;
+        this.wish = wish;
         id.setText(displayedIndex + ". ");
 
-        String descWithType = income.getDesc().fullDesc;
+        String descWithType = wish.getDesc().fullDesc;
         desc.setText(descWithType);
-        date.setText(income.getDate().toString());
-        amt.setText("$" + income.getAmount().value);
-        category.setText(income.getCategory().categoryName);
+        date.setText(wish.getDate().toString());
+        amt.setText("$" + wish.getAmount().value);
+        category.setText(wish.getCategory().toString());
 
-        income.getTags().stream()
+        wish.getTags().stream()
                 .sorted(Comparator.comparing(tag -> tag.tagName))
                 .forEach(tag -> tags.getChildren().add(new Label(tag.tagName)));
     }
@@ -65,13 +67,14 @@ public class IncomeCard extends UiPart<Region> {
         }
 
         // instanceof handles nulls
-        if (!(other instanceof IncomeCard)) {
+        if (!(other instanceof EntryCard)) {
             return false;
         }
 
         // state check
-        IncomeCard card = (IncomeCard) other;
+        WishListCard card = (WishListCard) other;
         return id.getText().equals(card.id.getText())
-                && income.equals(card.income);
+                && wish.equals(card.wish);
     }
 }
+
