+++ src/main/resources/view/autoexpense/AutoExpenseListCard.fxml
