+++ src/main/resources/view/condition/ConditionListCard.fxml
