+++ src/main/resources/view/condition/ConditionListPanel.fxml
