+++ src/main/resources/view/entry/EntryListCard.fxml
