+++ src/main/resources/view/entry/EntryListPanel.fxml
