+++ src/main/resources/view/reminder/WishReminderListCard.fxml
