+++ src/main/resources/view/wishlist/WishListCard.fxml
