+++ src/main/resources/view/wishlist/WishListPanel.fxml
